Adds the field 'lastUpdatedTime' to nse response
Self-explanatory, I was thinking of adding a separate function to convert the string into a proper datetime() object, is it better to convert to a datetime() object?.

The cleaner way you be to pass the whole (buffer) to clean_server_respone. But I just wanted a quick fix. It isn't even harmful.